api/requirements.txt
django
django-cors-headers
djangorestframework
psycopg2-binary
python-dotenv
sqlparse